+++ src/main/java/com/desafiotecnicoapi/apiVeiculos/service/ConsumoService.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 
 public interface ConsumoService {
     BigDecimal computeTotalPreco(BigDecimal totalPreco, BigDecimal combustivelConsumido);
-    BigDecimal computeUso(BigDecimal distancia, BigDecimal consumoMedio);
+    float computeUso(BigDecimal distancia, BigDecimal consumoMedio);
 }
 }
